+++ AlchemistCode/Cards/Basic/Jab.cs
@@ -13,6 +13,7 @@ public class Jab : AlchemistCard
     public Jab() : base(1, CardType.Attack, CardRarity.Basic, TargetType.AnyEnemy)
     {
         WithCalculatedDamage(6, static (card, _) => Dose(card), ValueProp.Move, 3);
+        WithKeyword(AlchemistKeywords.Laced);
         WithVar("Poison", 2, 1);
         WithTip(typeof(PoisonPower));
     }