+++ AlchemistCode/Cards/Rare/LashOut.cs
@@ -16,6 +16,7 @@ public class LashOut : AlchemistCard
     public LashOut() : base(1, CardType.Attack, CardRarity.Rare, TargetType.AnyEnemy)
     {
         WithCalculatedDamage(3, static (card, _) => Dose(card), ValueProp.Move, 2);
+        WithKeyword(AlchemistKeywords.Laced);
         WithTip(typeof(PoisonPower));
     }